--- OGL.rc
@@ -1,2 +1,0 @@
-#include "OGL.h"
-MYICON ICON kp.ico
