+++ src/lib/vmath.h
