+++ src/res/images/kp.ico
